other/python375_requirements.txt
absl-py==1.4.0
aiohttp==3.7.4.post0
altgraph==0.17
amiyabot==1.1.0
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
asgiref==3.3.4
astunparse==1.6.3
async-timeout==3.0.1
attrs==22.2.0
audioread==3.0.0
autopep8==2.0.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
binaryornot==0.4.4
blis==0.7.9
boto3==1.26.84
botocore==1.29.84
bs4==0.0.1
cachetools==5.3.0
catalogue==2.0.8
certifi==2020.6.20
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
ChatterBot==1.0.8
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
confection==0.0.4
config==0.5.1
construct==2.10.68
construct-typing==0.5.6
cookiecutter==2.3.0
cowsay==5.0
cryptography==3.4.6
cycler==0.10.0
cymem==2.0.7
databases==0.4.3
decorator==4.4.2
Deprecated==1.2.14
Django==3.2.4
dnspython==2.1.0
docopt==0.6.2
dominate==2.9.1
exceptiongroup==1.1.0
Faker==18.13.0
fastapi==0.79.1
fastgm==0.4.1
fire==0.4.0
Flask==1.1.2
flatbuffers==23.3.3
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.28
gmssl==3.2.2
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graiax-silkcoder==0.2.8
greenlet==1.1.2
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
hexdump==3.3
icmplib==3.0.4
icon-font-to-png==0.4.1
idna==2.10
image==1.5.33
ImageHash==4.3.1
imageio==2.21.3
imageio-ffmpeg==0.4.7
impacket==0.9.22
importlib==1.0.4
importlib-metadata==4.8.0
iniconfig==2.0.0
iso8601==2.0.0
itsdangerous==1.1.0
jaconv==0.3.4
jieba==0.42.1
Jinja2==2.11.3
jmespath==1.0.1
joblib==1.2.0
keras==2.11.0
kiwisolver==1.3.1
langcodes==3.3.0
lazy_loader==0.1
ldap3==2.9
ldapdomaindump==0.9.3
libclang==15.0.6.1
librosa==0.10.0.post2
llvmlite==0.39.1
lxml==4.6.3
lz4==4.3.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==1.1.1
mathparse==0.1.2
matplotlib==3.3.2
mdurl==0.1.2
more-bisect==0.1.1
moviepy==1.0.3
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
mysql==0.0.3
mysqlclient==2.0.3
Naked==0.1.31
netaddr==0.8.0
nltk==3.8.1
ntlm-auth==1.5.0
numba==0.56.4
numpy==1.21.2
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.0
palettable==3.3.0
paramiko==2.7.2
pathy==0.10.1
peewee==3.14.10
pefile==2021.5.24
Pillow==9.1.1
platformdirs==3.10.0
playsound==1.3.0
playwright==1.22.0
pluggy==1.0.0
pooch==1.6.0
preshed==3.0.8
proglog==0.1.10
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
py2exe==0.12.0.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.20
pycryptodome==3.15.0
pycryptodomex==3.18.0
pydantic==1.9.2
pydub==0.25.1
pyee==8.1.0
pygame==2.3.0
Pygments==2.16.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
pyminifier==2.1
pymongo==4.3.3
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypiwin32==223
pypng==0.20220715.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyserial==3.5
PySide2==5.15.2
PySocks==1.7.1
pytest==7.2.2
pytest-base-url==2.0.0
pytest-playwright==0.3.3
python-dateutil==2.8.1
python-lzo==1.15
python-slugify==8.0.1
pyttsx3==2.90
pytz==2021.1
PyWavelets==1.3.0
pywin32==301
pywin32-ctypes==0.2.0
PyYAML==6.0.1
qrcode==7.4.2
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
s3transfer==0.6.0
scapy==2.5.0
scikit-learn==1.0.2
scipy==1.7.0
shellescape==3.8.1
shiboken2==5.15.2
six==1.16.0
smart-open==6.3.0
smbprotocol==1.0.2.dev0
smmap==5.0.0
sniffio==1.3.0
snownlp==0.12.3
socks==0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.1
soxr==0.3.4
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.9.0
SQLAlchemy==1.3.24
sqlparse==0.4.1
srsly==2.4.5
starlette==0.19.1
stylecloud==0.5.2
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.0.1
text-unidecode==1.3
textblob==0.17.1
thinc==8.1.7
threadpoolctl==3.1.0
tinycss==0.4
tomli==2.0.1
tqdm==4.64.0
typer==0.7.0
typing_extensions==4.7.1
urllib3==1.26.3
uvicorn==0.18.3
vaderSentiment==3.3.2
wasabi==1.1.1
websockets==10.1
Werkzeug==1.0.1
wordcloud==1.8.2.2
wrapt==1.15.0
yarl==1.9.2
you-get==0.4.1700
zhon==1.1.5
zipp==3.5.0
